backend/requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
email-validator>=2.0.0

# HTTP & Environment
requests>=2.31.0
httpx>=0.25.0
python-dotenv>=1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0

# Linting & Formatting
ruff>=0.1.0
